pyproject.toml
# SPDX-FileCopyrightText: 2023 Phu Hung Nguyen <phuhnguyen@outlook.com>
# SPDX-License-Identifier: CC0-1.0

[tool.poetry]
name = "markdown-gettext"
version = "0.1.0"
description = "Markdown i18n with gettext"
authors = ["Phu Hung Nguyen <phuhnguyen@outlook.com>"]
license = "LGPL-2.1-or-later"
readme = "README.md"
repository = "https://github.com/phunh/markdown-gettext"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Internationalization",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3"
]
packages = [{include = "markdown_gettext"}]

[tool.poetry.dependencies]
python = "^3.8"
markdown-it-py = {extras = ["plugins"], version = "^3.0.0"}
polib = "^1.2.0"
Pygments = "^2.16.1"
PyYAML = "^6.0.1"

[tool.poetry.scripts]
md-gettext = 'markdown_gettext.cli:main'
markdown-gettext = 'markdown_gettext.cli:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
